zbus::blocking::connection

Struct Connection

pub struct Connection { /* private fields */ }

A blocking wrapper of zbus::Connection.

Most of the API is very similar to zbus::Connection, except it’s blocking.

Implementations

impl Connection

pub fn session() -> Result<Self>

Create a Connection to the session/user message bus.

pub fn system() -> Result<Self>

Create a Connection to the system-wide message bus.

pub fn max_queued(&self) -> usize

The capacity of the main (unfiltered) queue.

pub fn set_max_queued(self, max: usize)

Set the capacity of the main (unfiltered) queue.

pub fn server_guid(&self) -> &str

The server’s GUID.

pub fn unique_name(&self) -> Option<&OwnedUniqueName>

The unique name as assigned by the message bus or None if not a message bus connection.

pub fn send(&self, msg: &Message) -> Result<()>

Send msg to the peer.

pub fn call_method<'d, 'p, 'i, 'm, D, P, I, M, B>( &self, destination: Option<D>, path: P, iface: Option<I>, method_name: M, body: &B, ) -> Result<Message>
where D: TryInto<BusName<'d>>, P: TryInto<ObjectPath<'p>>, I: TryInto<InterfaceName<'i>>, M: TryInto<MemberName<'m>>, D::Error: Into<Error>, P::Error: Into<Error>, I::Error: Into<Error>, M::Error: Into<Error>, B: Serialize + DynamicType,

Send a method call.

Create a method-call message, send it over the connection, then wait for the reply.

On successful reply, an Ok(Message) is returned. On error, an Err is returned. D-Bus error replies are returned as Error::MethodError.

pub fn emit_signal<'d, 'p, 'i, 'm, D, P, I, M, B>( &self, destination: Option<D>, path: P, iface: I, signal_name: M, body: &B, ) -> Result<()>
where D: TryInto<BusName<'d>>, P: TryInto<ObjectPath<'p>>, I: TryInto<InterfaceName<'i>>, M: TryInto<MemberName<'m>>, D::Error: Into<Error>, P::Error: Into<Error>, I::Error: Into<Error>, M::Error: Into<Error>, B: Serialize + DynamicType,

Emit a signal.

Create a signal message, and send it over the connection.

pub fn reply<B>(&self, call: &Message, body: &B) -> Result<()>
where B: Serialize + DynamicType,

Reply to a message.

Given an existing message (likely a method call), send a reply back to the caller with the given body.

pub fn reply_error<'e, E, B>( &self, call: &Message, error_name: E, body: &B, ) -> Result<()>
where B: Serialize + DynamicType, E: TryInto<ErrorName<'e>>, E::Error: Into<Error>,

Reply an error to a message.

Given an existing message (likely a method call), send an error reply back to the caller with the given error_name and body.

Returns the message serial number.

pub fn reply_dbus_error( &self, call: &Header<'_>, err: impl DBusError, ) -> Result<()>

Reply to a method call with an error.

Given an existing method call message header, send an error reply back to the caller using one of the standard interface reply types.

Returns the message serial number.

pub fn request_name<'w, W>(&self, well_known_name: W) -> Result<()>
where W: TryInto<WellKnownName<'w>>, W::Error: Into<Error>,

Register a well-known name for this service on the bus.

Blocking version of crate::Connection::request_name. See docs there for more details and caveats.

pub fn request_name_with_flags<'w, W>( &self, well_known_name: W, flags: BitFlags<RequestNameFlags>, ) -> Result<RequestNameReply>
where W: TryInto<WellKnownName<'w>>, W::Error: Into<Error>,

Register a well-known name for this service on the bus.

Blocking version of crate::Connection::request_name_with_flags. See docs there for more details and caveats.

pub fn release_name<'w, W>(&self, well_known_name: W) -> Result<bool>
where W: TryInto<WellKnownName<'w>>, W::Error: Into<Error>,

Deregister a previously registered well-known name for this service on the bus.

Use this method to deregister a well-known name, registered through Connection::request_name.

Unless an error is encountered, returns Ok(true) if name was previously registered with the bus through self and it has now been successfully deregistered, Ok(false) if name was not previously registered or already deregistered.

pub fn is_bus(&self) -> bool

Checks if self is a connection to a message bus.

This will return false for p2p connections.

pub fn object_server(&self) -> impl Deref<Target = ObjectServer> + '_

Get a reference to the associated ObjectServer.

The ObjectServer is created on-demand.

pub fn inner(&self) -> &Connection

Get a reference to the underlying async Connection.

pub fn into_inner(self) -> Connection

Get the underlying async Connection, consuming self.

pub fn monitor_activity(&self) -> EventListener

Returns a listener, notified on various connection activity.

This function is meant for the caller to implement idle or timeout on inactivity.

pub fn peer_credentials(&self) -> Result<ConnectionCredentials>

Returns the peer credentials.

The fields are populated on the best effort basis. Some or all fields may not even make sense for certain sockets or on certain platforms and hence will be set to None.

Caveats

Currently unix_group_ids and linux_security_label fields are not populated.

pub fn close(self) -> Result<()>

Close the connection.

After this call, all reading and writing operations will fail.