zbus::blocking

Module object_server

The object server API.

Structs

InterfaceRef

Wrapper over an interface, along with its corresponding SignalContext instance. A reference to the underlying interface may be obtained via InterfaceRef::get and InterfaceRef::get_mut.

ObjectServer

A blocking wrapper of crate::ObjectServer.