zbus::blocking::proxy

Struct Proxy

pub struct Proxy<'a> { /* private fields */ }

A blocking wrapper of crate::Proxy.

This API is mostly the same as crate::Proxy, except that all its methods block to completion.

Example

use std::result::Result;
use std::error::Error;
use zbus::blocking::{Connection, Proxy};

fn main() -> Result<(), Box<dyn Error>> {
    let connection = Connection::session()?;
    let p = Proxy::new(
        &connection,
        "org.freedesktop.DBus",
        "/org/freedesktop/DBus",
        "org.freedesktop.DBus",
    )?;
    // owned return value
    let _id: String = p.call("GetId", &())?;
    // borrowed return value
    let body = p.call_method("GetId", &())?.body();
    let _id: &str = body.deserialize()?;
    Ok(())
}

Note

It is recommended to use the proxy macro, which provides a more convenient and type-safe façade Proxy derived from a Rust trait.

Current limitations:

At the moment, Proxy doesn’t prevent auto-launching.

Implementations

impl<'a> Proxy<'a>

pub fn new<D, P, I>( conn: &Connection, destination: D, path: P, interface: I, ) -> Result<Proxy<'a>>
where D: TryInto<BusName<'a>>, P: TryInto<ObjectPath<'a>>, I: TryInto<InterfaceName<'a>>, D::Error: Into<Error>, P::Error: Into<Error>, I::Error: Into<Error>,

Create a new Proxy for the given destination/path/interface.

pub fn new_owned<D, P, I>( conn: Connection, destination: D, path: P, interface: I, ) -> Result<Proxy<'a>>
where D: TryInto<BusName<'static>>, P: TryInto<ObjectPath<'static>>, I: TryInto<InterfaceName<'static>>, D::Error: Into<Error>, P::Error: Into<Error>, I::Error: Into<Error>,

Create a new Proxy for the given destination/path/interface, taking ownership of all passed arguments.

pub fn connection(&self) -> &Connection

Get a reference to the associated connection.

pub fn destination(&self) -> &BusName<'_>

Get a reference to the destination service name.

pub fn path(&self) -> &ObjectPath<'_>

Get a reference to the object path.

pub fn interface(&self) -> &InterfaceName<'_>

Get a reference to the interface.

pub fn introspect(&self) -> Result<String>

Introspect the associated object, and return the XML description.

See the xml module for parsing the result.

pub fn cached_property<T>(&self, property_name: &str) -> Result<Option<T>>
where T: TryFrom<OwnedValue>, T::Error: Into<Error>,

Get the cached value of the property property_name.

This returns None if the property is not in the cache. This could be because the cache was invalidated by an update, because caching was disabled for this property or proxy, or because the cache has not yet been populated. Use get_property to fetch the value from the peer.

pub fn cached_property_raw<'p>( &'p self, property_name: &'p str, ) -> Option<impl Deref<Target = Value<'static>> + 'p>

Get the cached value of the property property_name.

Same as cached_property, but gives you access to the raw value stored in the cache. This is useful if you want to avoid allocations and cloning.

pub fn get_property<T>(&self, property_name: &str) -> Result<T>
where T: TryFrom<OwnedValue>, T::Error: Into<Error>,

Get the property property_name.

Get the property value from the cache or call the Get method of the org.freedesktop.DBus.Properties interface.

pub fn set_property<'t, T>(&self, property_name: &str, value: T) -> Result<()>
where T: 't + Into<Value<'t>>,

Set the property property_name.

Effectively, call the Set method of the org.freedesktop.DBus.Properties interface.

pub fn call_method<'m, M, B>(&self, method_name: M, body: &B) -> Result<Message>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>, B: Serialize + DynamicType,

Call a method and return the reply.

Typically, you would want to use call method instead. Use this method if you need to deserialize the reply message manually (this way, you can avoid the memory allocation/copying, by deserializing the reply to an unowned type).

pub fn call<'m, M, B, R>(&self, method_name: M, body: &B) -> Result<R>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>, B: Serialize + DynamicType, R: for<'d> DynamicDeserialize<'d>,

Call a method and return the reply body.

Use call_method instead if you need to deserialize the reply manually/separately.

pub fn call_with_flags<'m, M, B, R>( &self, method_name: M, flags: BitFlags<MethodFlags>, body: &B, ) -> Result<Option<R>>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>, B: Serialize + DynamicType, R: for<'d> DynamicDeserialize<'d>,

Call a method and return the reply body, optionally supplying a set of method flags to control the way the method call message is sent and handled.

Use call instead if you do not need any special handling via additional flags. If the NoReplyExpected flag is passed , this will return None immediately after sending the message, similar to call_noreply

pub fn call_noreply<'m, M, B>(&self, method_name: M, body: &B) -> Result<()>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>, B: Serialize + DynamicType,

Call a method without expecting a reply

This sets the NoReplyExpected flag on the calling message and does not wait for a reply.

pub fn receive_signal<'m, M>( &self, signal_name: M, ) -> Result<SignalIterator<'m>>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>,

Create a stream for signal named signal_name.

Errors

Apart from general I/O errors that can result from socket communications, calling this method will also result in an error if the destination service has not yet registered its well-known name with the bus (assuming you’re using the well-known name as destination).

pub fn receive_signal_with_args<'m, M>( &self, signal_name: M, args: &[(u8, &str)], ) -> Result<SignalIterator<'m>>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>,

Same as Proxy::receive_signal but with a filter.

The D-Bus specification allows you to filter signals by their arguments, which helps avoid a lot of unnecessary traffic and processing since the filter is run on the server side. Use this method where possible. Note that this filtering is limited to arguments of string types.

The arguments are passed as a tuples of argument index and expected value.

pub fn receive_all_signals(&self) -> Result<SignalIterator<'static>>

Create a stream for all signals emitted by this service.

Errors

Apart from general I/O errors that can result from socket communications, calling this method will also result in an error if the destination service has not yet registered its well-known name with the bus (assuming you’re using the well-known name as destination).

pub fn receive_property_changed<'name: 'a, T>( &self, name: &'name str, ) -> PropertyIterator<'a, T>

Get an iterator to receive owner changed events.

If the proxy destination is a unique name, the stream will be notified of the peer disconnection from the bus (with a None value).

If the proxy destination is a well-known name, the stream will be notified whenever the name owner is changed, either by a new peer being granted ownership (Some value) or when the name is released (with a None value).

Note that zbus doesn’t queue the updates. If the listener is slower than the receiver, it will only receive the last update.

pub fn receive_owner_changed(&self) -> Result<OwnerChangedIterator<'_>>

Get an iterator to receive property changed events.

Note that zbus doesn’t queue the updates. If the listener is slower than the receiver, it will only receive the last update.

pub fn inner(&self) -> &Proxy<'a>

Get a reference to the underlying async Proxy.

pub fn into_inner(self) -> Proxy<'a>

Get the underlying async Proxy, consuming self.