zbus

Enum Error

#[non_exhaustive]
pub enum Error {

    InterfaceNotFound,
    Address(String),
    InputOutput(Arc<Error>),
    InvalidField,
    ExcessData,
    Variant(Error),
    Names(Error),
    IncorrectEndian,
    Handshake(String),
    InvalidReply,
    MethodError(OwnedErrorName, Option<String>, Message),
    MissingField,
    InvalidGUID,
    Unsupported,
    FDO(Box<Error>),
    NameTaken,
    InvalidMatchRule,
    Failure(String),
    MissingParameter(&'static str),
    InvalidSerial,
    InterfaceExists(InterfaceName<'static>, ObjectPath<'static>),

}

The error type for zbus.

The various errors that can be reported by this crate.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InterfaceNotFound

Interface not found

Address(String)

Invalid D-Bus address.

InputOutput(Arc<Error>)

An I/O error.

InvalidField

Invalid message field.

ExcessData

Data too large.

Variant(Error)

A zvariant error.

Names(Error)

A zbus_names error.

IncorrectEndian

Endian signature invalid or doesnโt match expectation.

Handshake(String)

Initial handshake error.

InvalidReply

Unexpected or incorrect reply.

MethodError(OwnedErrorName, Option<String>, Message)

A D-Bus method error reply.

MissingField

A required field is missing in the message headers.

InvalidGUID

Invalid D-Bus GUID.

Unsupported

Unsupported function, or support currently lacking.

FDO(Box<Error>)

A fdo::Error transformed into Error.

NameTaken

The requested name was already claimed by another peer.

InvalidMatchRule

Invalid match rule string.

Failure(String)

Generic error.

MissingParameter(&'static str)

A required parameter was missing.

InvalidSerial

Serial number in the message header is 0 (which is invalid).

InterfaceExists(InterfaceName<'static>, ObjectPath<'static>)

The given interface already exists at the given path.