Enum zbus::MethodFlags

#[repr(u8)]
pub enum MethodFlags {
    NoReplyExpected,
    NoAutoStart,
    AllowInteractiveAuth,
}

Flags to use with Proxy::call_with_flags.

Variants

NoReplyExpected

No response is expected from this method call, regardless of whether the signature for the interface method indicates a reply type. When passed, call_with_flags will return Ok(None) immediately after successfully sending the method call.

Errors encountered while making the call will still be returned as an Err variant, but any errors that are triggered by the receiver’s handling of the call will not be delivered.

NoAutoStart

When set on a call whose destination is a message bus, this flag will instruct the bus not to launch a service to handle the call if no application on the bus owns the requested name.

This flag is ignored when using a peer-to-peer connection.

AllowInteractiveAuth

Indicates to the receiver that this client is prepared to wait for interactive authorization, which might take a considerable time to complete. For example, the receiver may query the user for confirmation via polkit or a similar framework.