zbus::fdo

Module dbus

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

ConnectionCredentials

Credentials of a process connected to a bus server.

DBusProxy

Proxy for the org.freedesktop.DBus interface.

NameAcquired

A NameAcquired signal.

NameAcquiredArgs

NameAcquired signal arguments.

NameAcquiredStream

A Stream implementation that yields NameAcquired signals.

NameLost

A NameLost signal.

NameLostArgs

NameLost signal arguments.

NameLostStream

A Stream implementation that yields NameLost signals.

NameOwnerChanged

A NameOwnerChanged signal.

NameOwnerChangedArgs

NameOwnerChanged signal arguments.

NameOwnerChangedStream

A Stream implementation that yields NameOwnerChanged signals.

Enums

ReleaseNameReply

The return code of the DBusProxy::release_name method.

RequestNameFlags

The flags used by the DBusProxy::request_name method.

RequestNameReply

The return code of the DBusProxy::request_name method.

StartServiceReply

The return code of the DBusProxy::start_service_by_name method.