zbus::fdo

Enum RequestNameFlags

#[repr(u32)]
pub enum RequestNameFlags {
    AllowReplacement = 1,
    ReplaceExisting = 2,
    DoNotQueue = 4,
}

The flags used by the bus request_name method.

Variants

AllowReplacement = 1

If an application A specifies this flag and succeeds in becoming the owner of the name, and another application B later calls request_name with the ReplaceExisting flag, then application A will lose ownership and receive a org.freedesktop.DBus.NameLost signal, and application B will become the new owner. If AllowReplacement is not specified by application A, or ReplaceExisting is not specified by application B, then application B will not replace application A as the owner.

ReplaceExisting = 2

Try to replace the current owner if there is one. If this flag is not set the application will only become the owner of the name if there is no current owner. If this flag is set, the application will replace the current owner if the current owner specified AllowReplacement.

DoNotQueue = 4

Without this flag, if an application requests a name that is already owned, the application will be placed in a queue to own the name when the current owner gives it up. If this flag is given, the application will not be placed in the queue, the request for the name will simply fail. This flag also affects behavior when an application is replaced as name owner; by default the application moves back into the waiting queue, unless this flag was provided when the application became the name owner.