zbus::fdo

Module introspectable

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

Introspectable 🔒

Service-side implementation for the org.freedesktop.DBus.Introspectable interface. This interface is implemented automatically for any object registered to the ObjectServer.

IntrospectableProxy

Proxy for the org.freedesktop.DBus.Introspectable interface.