zbus::fdo

Module properties

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

Properties

Service-side implementation for the org.freedesktop.DBus.Properties interface. This interface is implemented automatically for any object registered to the ObjectServer.

PropertiesChanged

A PropertiesChanged signal.

PropertiesChangedArgs

PropertiesChanged signal arguments.

PropertiesChangedStream

A Stream implementation that yields PropertiesChanged signals.

PropertiesProxy

Proxy for the org.freedesktop.DBus.Properties interface.

Traits

PropertiesSignals

Trait providing all signal emission methods for Properties.