zbus::fdo

Struct DBusProxy

pub struct DBusProxy<'p>(/* private fields */);

Proxy for the org.freedesktop.DBus interface.

Implementations

impl<'p> DBusProxy<'p>

pub async fn new(conn: &Connection) -> Result<DBusProxy<'p>>

Creates a new proxy with the default service and path.

pub fn builder(conn: &Connection) -> Builder<'p, Self>

Returns a customizable builder for this proxy.

pub fn into_inner(self) -> Proxy<'p>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'p>

The reference to the underlying zbus::Proxy.

pub fn inner_mut(&mut self) -> &mut Proxy<'p>

The mutable reference to the underlying zbus::Proxy.

pub async fn add_match_rule(&self, rule: MatchRule<'_>) -> Result<()>

Adds a match rule to match messages going through the message bus

pub async fn get_adt_audit_session_data( &self, bus_name: BusName<'_>, ) -> Result<Vec<u8>>

Returns auditing data used by Solaris ADT, in an unspecified binary format.

pub async fn get_connection_credentials( &self, bus_name: BusName<'_>, ) -> Result<ConnectionCredentials>

Returns as many credentials as possible for the process connected to the server.

pub async fn get_connection_selinux_security_context( &self, bus_name: BusName<'_>, ) -> Result<Vec<u8>>

Returns the security context used by SELinux, in an unspecified format.

pub async fn get_connection_unix_process_id( &self, bus_name: BusName<'_>, ) -> Result<u32>

Returns the Unix process ID of the process connected to the server.

pub async fn get_connection_unix_user( &self, bus_name: BusName<'_>, ) -> Result<u32>

Returns the Unix user ID of the process connected to the server.

pub async fn get_id(&self) -> Result<OwnedGuid>

Gets the unique ID of the bus.

pub async fn get_name_owner(&self, name: BusName<'_>) -> Result<OwnedUniqueName>

Returns the unique connection name of the primary owner of the name given.

pub async fn hello(&self) -> Result<OwnedUniqueName>

Returns the unique name assigned to the connection.

pub async fn list_activatable_names(&self) -> Result<Vec<OwnedBusName>>

Returns a list of all names that can be activated on the bus.

pub async fn list_names(&self) -> Result<Vec<OwnedBusName>>

Returns a list of all currently-owned names on the bus.

pub async fn list_queued_owners( &self, name: WellKnownName<'_>, ) -> Result<Vec<OwnedUniqueName>>

List the connections currently queued for a bus name.

pub async fn name_has_owner(&self, name: BusName<'_>) -> Result<bool>

Checks if the specified name exists (currently has an owner).

pub async fn release_name( &self, name: WellKnownName<'_>, ) -> Result<ReleaseNameReply>

Ask the message bus to release the method caller’s claim to the given name.

pub async fn reload_config(&self) -> Result<()>

Reload server configuration.

pub async fn remove_match_rule(&self, rule: MatchRule<'_>) -> Result<()>

Removes the first rule that matches.

pub async fn request_name( &self, name: WellKnownName<'_>, flags: BitFlags<RequestNameFlags>, ) -> Result<RequestNameReply>

Ask the message bus to assign the given name to the method caller.

pub async fn start_service_by_name( &self, name: WellKnownName<'_>, flags: u32, ) -> Result<u32>

Tries to launch the executable associated with a name (service activation), as an explicit request.

pub async fn update_activation_environment( &self, environment: HashMap<&str, &str>, ) -> Result<()>

This method adds to or modifies that environment when activating services.

pub async fn receive_name_owner_changed( &self, ) -> Result<NameOwnerChangedStream<'static>>

Create a stream that receives NameOwnerChanged signals.

This a convenient wrapper around zbus::Proxy::receive_signal. This signal indicates that the owner of a name has changed. It’s also the signal to use to detect the appearance of new names on the bus.

pub async fn receive_name_owner_changed_with_args( &self, args: &[(u8, &str)], ) -> Result<NameOwnerChangedStream<'static>>

Create a stream that receives NameOwnerChanged signals.

This a convenient wrapper around zbus::Proxy::receive_signal_with_args. This signal indicates that the owner of a name has changed. It’s also the signal to use to detect the appearance of new names on the bus.

pub async fn receive_name_lost(&self) -> Result<NameLostStream<'static>>

Create a stream that receives NameLost signals.

This a convenient wrapper around zbus::Proxy::receive_signal. This signal is sent to a specific application when it loses ownership of a name.

pub async fn receive_name_lost_with_args( &self, args: &[(u8, &str)], ) -> Result<NameLostStream<'static>>

Create a stream that receives NameLost signals.

This a convenient wrapper around zbus::Proxy::receive_signal_with_args. This signal is sent to a specific application when it loses ownership of a name.

pub async fn receive_name_acquired(&self) -> Result<NameAcquiredStream<'static>>

Create a stream that receives NameAcquired signals.

This a convenient wrapper around zbus::Proxy::receive_signal. This signal is sent to a specific application when it gains ownership of a name.

pub async fn receive_name_acquired_with_args( &self, args: &[(u8, &str)], ) -> Result<NameAcquiredStream<'static>>

Create a stream that receives NameAcquired signals.

This a convenient wrapper around zbus::Proxy::receive_signal_with_args. This signal is sent to a specific application when it gains ownership of a name.

pub async fn features(&self) -> Result<Vec<String>>

This property lists abstract “features” provided by the message bus, and can be used by clients to detect the capabilities of the message bus with which they are communicating.

pub fn cached_features( &self, ) -> Result<Option<<Result<Vec<String>> as ResultAdapter>::Ok>, <Result<Vec<String>> as ResultAdapter>::Err>

Get the cached value of the Features property, or None if the property is not cached.

pub async fn receive_features_changed( &self, ) -> PropertyStream<'p, <Result<Vec<String>> as ResultAdapter>::Ok>

Create a stream for the Features property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.

pub async fn interfaces(&self) -> Result<Vec<OwnedInterfaceName>>

This property lists interfaces provided by the /org/freedesktop/DBus object, and can be used by clients to detect the capabilities of the message bus with which they are communicating. Unlike the standard Introspectable interface, querying this property does not require parsing XML. This property was added in version 1.11.x of the reference implementation of the message bus.

The standard org.freedesktop.DBus and org.freedesktop.DBus.Properties interfaces are not included in the value of this property, because their presence can be inferred from the fact that a method call on org.freedesktop.DBus.Properties asking for properties of org.freedesktop.DBus was successful. The standard org.freedesktop.DBus.Peer and org.freedesktop.DBus.Introspectable interfaces are not included in the value of this property either, because they do not indicate features of the message bus implementation.

pub fn cached_interfaces( &self, ) -> Result<Option<<Result<Vec<OwnedInterfaceName>> as ResultAdapter>::Ok>, <Result<Vec<OwnedInterfaceName>> as ResultAdapter>::Err>

Get the cached value of the Interfaces property, or None if the property is not cached.

pub async fn receive_interfaces_changed( &self, ) -> PropertyStream<'p, <Result<Vec<OwnedInterfaceName>> as ResultAdapter>::Ok>

Create a stream for the Interfaces property changes. This is a convenient wrapper around zbus::Proxy::receive_property_changed.