zbus::fdo

Struct ObjectManager

pub struct ObjectManager;

Service-side Object Manager interface implementation.

The recommended path to add this interface at is the path form of the well-known name of a D-Bus service, or below. For example, if a D-Bus service is available at the well-known name net.example.ExampleService1, this interface should typically be registered at /net/example/ExampleService1, or below (to allow for multiple object managers in a service).

It is supported, but not recommended, to add this interface at the root path, /.

When added to an ObjectServer, InterfacesAdded signal is emitted for all the objects under the path its added at. You can use this fact to minimize the signal emissions by populating the entire (sub)tree under path before registering an object manager.

Implementations

impl ObjectManager

pub async fn interfaces_added( ctxt: &SignalContext<'_>, object_path: &ObjectPath<'_>, interfaces_and_properties: &HashMap<InterfaceName<'_>, HashMap<&str, Value<'_>>>, ) -> Result<()>

This signal is emitted when either a new object is added or when an existing object gains one or more interfaces. The interfaces_and_properties argument contains a map with the interfaces and properties (if any) that have been added to the given object path.

pub async fn interfaces_removed( ctxt: &SignalContext<'_>, object_path: &ObjectPath<'_>, interfaces: &[InterfaceName<'_>], ) -> Result<()>

This signal is emitted whenever an object is removed or it loses one or more interfaces. The interfaces parameters contains a list of the interfaces that were removed.