Struct zbus::fdo::StatsProxy

pub struct StatsProxy<'c>(_);

Proxy for the org.freedesktop.DBus.Stats interface.

Implementations

impl<'c> StatsProxy<'c>

pub fn new(conn: &Connection) -> Result<Self>

Creates a new proxy with the default service & path.

pub fn new_for(
conn: &Connection,
destination: &'c str,
path: &'c str
) -> Result<Self>

Creates a new proxy for the given destination and path.

pub fn new_for_owned(
conn: Connection,
destination: String,
path: String
) -> Result<Self>

Same as new_for but takes ownership of the passed arguments.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy

The reference to the underlying zbus::Proxy.

pub fn get_stats(&self) -> Result<Vec<HashMap<String, OwnedValue>>>

GetStats (undocumented)

pub fn get_connection_stats(
&self,
n1: &str
) -> Result<Vec<HashMap<String, OwnedValue>>>

GetConnectionStats (undocumented)

pub fn get_all_match_rules(&self) -> Result<Vec<HashMap<String, Vec<String>>>>

GetAllMatchRules (undocumented)

pub fn introspect(&self) -> Result<String>