zbus::match_rule

Struct Builder

pub struct Builder<'m>(/* private fields */);

Builder for MatchRule.

This is created by MatchRule::builder.

Implementations

impl<'m> Builder<'m>

pub fn build(self) -> MatchRule<'m>

Build the MatchRule.

pub fn sender<B>(self, sender: B) -> Result<Self>
where B: TryInto<BusName<'m>>, B::Error: Into<Error>,

Set the sender.

pub fn msg_type(self, msg_type: Type) -> Self

Set the message type.

pub fn interface<I>(self, interface: I) -> Result<Self>
where I: TryInto<InterfaceName<'m>>, I::Error: Into<Error>,

Set the interface.

pub fn member<M>(self, member: M) -> Result<Self>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>,

Set the member.

pub fn path<P>(self, path: P) -> Result<Self>
where P: TryInto<ObjectPath<'m>>, P::Error: Into<Error>,

Set the path.

Note: Since both a path and a path namespace are not allowed to appear in a match rule at the same time, this overrides any path namespace previously set.

pub fn path_namespace<P>(self, path_namespace: P) -> Result<Self>
where P: TryInto<ObjectPath<'m>>, P::Error: Into<Error>,

Set the path namespace.

Note: Since both a path and a path namespace are not allowed to appear in a match rule at the same time, this overrides any path previously set.

pub fn destination<B>(self, destination: B) -> Result<Self>
where B: TryInto<UniqueName<'m>>, B::Error: Into<Error>,

Set the destination.

pub fn add_arg<S>(self, arg: S) -> Result<Self>
where S: Into<Str<'m>>,

Append an arguments.

Use this in instead of Builder::arg if you want to sequentially add args.

Errors

Error::InvalidMatchRule on attempt to add the 65th argument.

pub fn arg<S>(self, idx: u8, arg: S) -> Result<Self>
where S: Into<Str<'m>>,

Add an argument of a specified index.

Errors

Error::InvalidMatchRule if idx is greater than 64.

pub fn add_arg_path<P>(self, arg_path: P) -> Result<Self>
where P: TryInto<ObjectPath<'m>>, P::Error: Into<Error>,

Append a path argument.

Use this in instead of Builder::arg_path if you want to sequentially add args.

Errors

Error::InvalidMatchRule on attempt to add the 65th path argument.

pub fn arg_path<P>(self, idx: u8, arg_path: P) -> Result<Self>
where P: TryInto<ObjectPath<'m>>, P::Error: Into<Error>,

Add a path argument of a specified index.

Errors

Error::InvalidMatchRule if idx is greater than 64.

pub fn arg0ns<S>(self, namespace: S) -> Result<Self>
where S: Into<Str<'m>>,

Set 0th argument’s namespace.

The namespace be a valid bus name or a valid element of a bus name. For more information, see the spec.

Examples

// Valid namespaces
MatchRule::builder().arg0ns("org.mpris.MediaPlayer2").unwrap();
MatchRule::builder().arg0ns("org").unwrap();
MatchRule::builder().arg0ns(":org").unwrap();
MatchRule::builder().arg0ns(":1org").unwrap();

// Invalid namespaces
MatchRule::builder().arg0ns("org.").unwrap_err();
MatchRule::builder().arg0ns(".org").unwrap_err();
MatchRule::builder().arg0ns("1org").unwrap_err();
MatchRule::builder().arg0ns(".").unwrap_err();
MatchRule::builder().arg0ns("org..freedesktop").unwrap_err();