zbus
Module message

D-Bus Message.

Structs

Body

The body of a message.

Builder

A builder for Message

Header

The message header, containing all the metadata about the message.

Message

A D-Bus Message.

PrimaryHeader

The primary message header, which is present in all D-Bus messages.

Sequence

A position in the stream of Message objects received by a single zbus::Connection.

Enums

EndianSig

D-Bus code for endianness.

Flags

Pre-defined flags that can be passed in Message header.

Type

Message header representing the D-Bus type of the message.

Constants

NATIVE_ENDIAN_SIG

Signature of the target’s native endian.