zbus::message

Module header

Structs

Header

The message header, containing all the metadata about the message.

PrimaryHeader

The primary message header, which is present in all D-Bus messages.

Enums

EndianSig

D-Bus code for endianness.

Flags

Pre-defined flags that can be passed in message headers.

Type

Message header representing the D-Bus type of the message.

Constants

MAX_MESSAGE_SIZE 🔒

MIN_MESSAGE_SIZE 🔒

NATIVE_ENDIAN_SIG

Signature of the target’s native endian.

PRIMARY_HEADER_SIZE 🔒

Statics

SERIAL_NUM 🔒