zbus::message

Struct Builder

pub struct Builder<'a> { /* private fields */ }

A builder for Message

Implementations

impl<'a> Builder<'a>

pub fn method_call<'p: 'a, 'm: 'a, P, M>( path: P, method_name: M, ) -> Result<Self>
where P: TryInto<ObjectPath<'p>>, M: TryInto<MemberName<'m>>, P::Error: Into<Error>, M::Error: Into<Error>,

๐Deprecated since 4.0.0: Please use Message::method instead

Create a message of type Type::MethodCall.

pub fn signal<'p: 'a, 'i: 'a, 'm: 'a, P, I, M>( path: P, interface: I, name: M, ) -> Result<Self>
where P: TryInto<ObjectPath<'p>>, I: TryInto<InterfaceName<'i>>, M: TryInto<MemberName<'m>>, P::Error: Into<Error>, I::Error: Into<Error>, M::Error: Into<Error>,

๐Deprecated since 4.0.0: Please use Message::signal instead

Create a message of type Type::Signal.

pub fn method_return(reply_to: &Header<'_>) -> Result<Self>

๐Deprecated since 4.0.0: Please use Message::method_reply instead

Create a message of type Type::MethodReturn.

pub fn error<'e: 'a, E>(reply_to: &Header<'_>, name: E) -> Result<Self>
where E: TryInto<ErrorName<'e>>, E::Error: Into<Error>,

๐Deprecated since 4.0.0: Please use Message::method_error instead

Create a message of type Type::Error.

pub fn with_flags(self, flag: Flags) -> Result<Self>

Add flags to the message.

See Flags documentation for the meaning of the flags.

The function will return an error if invalid flags are given for the message type.

pub fn sender<'s: 'a, S>(self, sender: S) -> Result<Self>
where S: TryInto<UniqueName<'s>>, S::Error: Into<Error>,

Set the unique name of the sending connection.

pub fn path<'p: 'a, P>(self, path: P) -> Result<Self>
where P: TryInto<ObjectPath<'p>>, P::Error: Into<Error>,

Set the object to send a call to, or the object a signal is emitted from.

pub fn interface<'i: 'a, I>(self, interface: I) -> Result<Self>
where I: TryInto<InterfaceName<'i>>, I::Error: Into<Error>,

Set the interface to invoke a method call on, or that a signal is emitted from.

pub fn member<'m: 'a, M>(self, member: M) -> Result<Self>
where M: TryInto<MemberName<'m>>, M::Error: Into<Error>,

Set the member, either the method name or signal name.

pub fn destination<'d: 'a, D>(self, destination: D) -> Result<Self>
where D: TryInto<BusName<'d>>, D::Error: Into<Error>,

Set the name of the connection this message is intended for.

pub fn endian(self, endian: Endian) -> Self

Set the endianness of the message.

The default endianness is native.

pub fn build<B>(self, body: &B) -> Result<Message>
where B: Serialize + DynamicType,

Build the Message with the given body.

You may pass () as the body if the message has no body.

The caller is currently required to ensure that the resulting message contains the headers as compliant with the specification. Additional checks may be added to this builder over time as needed.

pub unsafe fn build_raw_body<'b, S>( self, body_bytes: &[u8], signature: S, fds: Vec<OwnedFd>, ) -> Result<Message>
where S: TryInto<Signature<'b>>, S::Error: Into<Error>,

Create a new message from a raw slice of bytes to populate the body with, rather than by serializing a value. The message body will be the exact bytes.

Safety

This method is unsafe because it can be used to build an invalid message.