zbus::message

Struct Header

pub struct Header<'m> { /* private fields */ }

The message header, containing all the metadata about the message.

This includes both the PrimaryHeader and Fields.

Implementations

impl<'m> Header<'m>

pub fn primary(&self) -> &PrimaryHeader

Get a reference to the primary header.

pub fn primary_mut(&mut self) -> &mut PrimaryHeader

Get a mutable reference to the primary header.

pub fn into_primary(self) -> PrimaryHeader

Get the primary header, consuming self.

pub fn message_type(&self) -> Type

The message type

pub fn path(&self) -> Option<&ObjectPath<'m>>

The object to send a call to, or the object a signal is emitted from.

pub fn interface(&self) -> Option<&InterfaceName<'m>>

The interface to invoke a method call on, or that a signal is emitted from.

pub fn member(&self) -> Option<&MemberName<'m>>

The member, either the method name or signal name.

pub fn error_name(&self) -> Option<&ErrorName<'m>>

The name of the error that occurred, for errors.

pub fn reply_serial(&self) -> Option<NonZeroU32>

The serial number of the message this message is a reply to.

pub fn destination(&self) -> Option<&BusName<'m>>

The name of the connection this message is intended for.

pub fn sender(&self) -> Option<&UniqueName<'m>>

Unique name of the sending connection.

pub fn signature(&self) -> Option<&Signature<'m>>

The signature of the message body.

pub fn unix_fds(&self) -> Option<u32>

The number of Unix file descriptors that accompany the message.