zbus

Module object_server

The object server API.

Modules

dispatch_notifier 🔒

The object server API.

interface 🔒

node 🔒

The object server API.

signal_emitter 🔒

Structs

InterfaceDeref

Opaque structure that derefs to an Interface type.

InterfaceDerefMut

Opaque structure that mutably derefs to an Interface type.

InterfaceRef

Wrapper over an interface, along with its corresponding SignalEmitter instance. A reference to the underlying interface may be obtained via InterfaceRef::get and InterfaceRef::get_mut.

ObjectServer

An object server, holding server-side D-Bus objects & interfaces.

ResponseDispatchNotifier

A response wrapper that notifies after the response has been sent.

SignalEmitter

A signal emitter.

Enums

DispatchResultDeprecated

A helper type returned by Interface callbacks.

DispatchResult2

A helper type returned by Interface callbacks.

Traits

Interface

This trait is used to dispatch messages to an interface instance.

Type Aliases

SignalContextDeprecated