zbus::object_server

Struct SignalEmitter

pub struct SignalEmitter<'s> {
    conn: Connection,
    path: ObjectPath<'s>,
    destination: Option<BusName<'s>>,
}

A signal emitter.

For signal emission using the high-level API, you’ll need instances of this type.

See crate::object_server::InterfaceRef::signal_emitter and crate::interface documentation for details and examples of this type in use.

Fields
conn: Connectionpath: ObjectPath<'s>destination: Option<BusName<'s>>
Implementations

impl<'s> SignalEmitter<'s>

pub fn new<P>(conn: &Connection, path: P) -> Result<Self>
where P: TryInto<ObjectPath<'s>>, P::Error: Into<Error>,

Create a new signal context for the given connection and object path.

pub fn from_parts(conn: Connection, path: ObjectPath<'s>) -> Self

Create a new signal context for the given connection and object path.

pub async fn emit<'i, 'm, I, M, B>( &self, interface: I, signal_name: M, body: &B, ) -> Result<()>
where I: TryInto<InterfaceName<'i>>, I::Error: Into<Error>, M: TryInto<MemberName<'m>>, M::Error: Into<Error>, B: Serialize + DynamicType,

Emit a signal on the given interface with the given signal name and body.

pub fn set_destination(self, destination: BusName<'s>) -> Self

Set the destination for the signal emission.

Signals are typically broadcasted and thus don’t have a destination. However, there are cases where you need to unicast signals to specific peers. This method allows you to set the destination for the signals emitted with this context.

pub fn connection(&self) -> &Connection

Get a reference to the associated connection.

pub fn path(&self) -> &ObjectPath<'s>

Get a reference to the associated object path.

pub fn destination(&self) -> Option<&BusName<'s>>

Get a reference to the associated destination (if any).

pub fn to_owned(&self) -> SignalEmitter<'static>

Create an owned clone of self.

pub fn into_owned(self) -> SignalEmitter<'static>

Convert into an owned clone of self.