zbus

Module proxy

The client-side proxy API.

Modules

builder 🔒

defaults 🔒

Structs

Builder

Builder for proxies.

OwnerChangedStream

A stream::Stream implementation that yields UniqueName when the bus owner changes.

PropertiesCache 🔒

PropertyChanged

A property changed event.

PropertyStream

A stream::Stream implementation that yields property change notifications.

PropertyValue 🔒

Proxy

A client-side interface proxy.

ProxyInner 🔒

ProxyInnerStatic 🔒

This is required to avoid having the Drop impl extend the lifetime ’a, which breaks zbus_xmlgen (and possibly other crates).

SignalStream

A stream::Stream implementation that yields signal messages.

Enums

CacheProperties

The properties caching mode.

CachingResult 🔒

Either 🔒

MethodFlags

Flags to use with Proxy::call_with_flags.

Constants

MAX_NAME_OWNER_CHANGED_SIGNALS_QUEUED 🔒

Traits

Defaults

Trait for the default associated values of a proxy.

ProxyImpl

This trait is implemented by all async proxies, which are generated with the proxy macro.

Type Aliases

OwnerChangedStreamMap 🔒