zbus::proxy

Trait ProxyDefault

pub trait ProxyDefault {
    const INTERFACE: Option<&'static str>;
    const DESTINATION: Option<&'static str>;
    const PATH: Option<&'static str>;
}

Trait for the default associated values of a proxy.

The trait is automatically implemented by the dbus_proxy macro on your behalf, and may be later used to retrieve the associated constants.

Required Associated Constants

const INTERFACE: Option<&'static str>

const DESTINATION: Option<&'static str>

const PATH: Option<&'static str>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.