Module zbus::quick_xml

Introspection XML support (quick-xml feature)

Thanks to the org.freedesktop.DBus.Introspectable interface, objects may be introspected at runtime, returning an XML string that describes the object.

This optional quick_xml module provides facilities to parse the XML data into more convenient Rust structures. The XML string may be parsed to a tree with Node.from_reader().

This module has a more type-safe API and uses a maintained XML parser. It will eventually replace the xml module. See also:

Introspection format in the DBus specification

Structs

Annotation

Annotations are generic key/value pairs of metadata.

Arg

An argument

Interface

An interface

Method

A method

Node

An introspection tree node (typically the root of the XML document).

Property

A property

Signal

A signal

Enums

ArgDirection

A direction of an argument

PropertyAccess

The possible property access types