Struct zbus::MessageHeader

pub struct MessageHeader<'m> { /* fields omitted */ }

The message header, containing all the metadata about the message.

This includes both the MessagePrimaryHeader and MessageFields.

Implementations

impl<'m> MessageHeader<'m>

pub fn new(primary: MessagePrimaryHeader, fields: MessageFields<'m>) -> Self

Create a new MessageHeader instance.

pub fn primary(&self) -> &MessagePrimaryHeader

Get a reference to the primary header.

pub fn primary_mut(&mut self) -> &mut MessagePrimaryHeader

Get a mutable reference to the primary header.

pub fn into_primary(self) -> MessagePrimaryHeader

Get the primary header, consuming self.

pub fn fields<'s>(&'s self) -> &'s MessageFields<'m>

Get a reference to the message fields.

pub fn fields_mut<'s>(&'s mut self) -> &'s mut MessageFields<'m>

Get a mutable reference to the message fields.

pub fn into_fields(self) -> MessageFields<'m>

Get the message fields, consuming self.

pub fn message_type(&self) -> Result<MessageType, MessageError>

The message type

pub fn path(&self) -> Result<Option<&ObjectPath>, MessageError>

The object to send a call to, or the object a signal is emitted from.

pub fn interface<'s>(&'s self) -> Result<Option<&'s str>, MessageError>

The interface to invoke a method call on, or that a signal is emitted from.

pub fn member<'s>(&'s self) -> Result<Option<&'s str>, MessageError>

The member, either the method name or signal name.

pub fn error_name<'s>(&'s self) -> Result<Option<&'s str>, MessageError>

The name of the error that occurred, for errors.

pub fn reply_serial(&self) -> Result<Option<u32>, MessageError>

The serial number of the message this message is a reply to.

pub fn destination<'s>(&'s self) -> Result<Option<&'s str>, MessageError>

The name of the connection this message is intended for.

pub fn sender<'s>(&'s self) -> Result<Option<&'s str>, MessageError>

Unique name of the sending connection.

pub fn signature(&self) -> Result<Option<&Signature>, MessageError>

The signature of the message body.

pub fn unix_fds(&self) -> Result<Option<u32>, MessageError>

The number of Unix file descriptors that accompany the message.