Trait zbus::ProxyDefault

pub trait ProxyDefault {
    const INTERFACE: &'static str;
    const DESTINATION: &'static str;
    const PATH: &'static str;
}

Trait for the default associated values of a proxy.

The trait is automatically implemented by the dbus_proxy macro on your behalf, and may be later used to retrieve the associated constants.

Required Associated Constants

const INTERFACE: &'static str

const DESTINATION: &'static str

const PATH: &'static str

Object Safety

This trait is not object safe.