zbus

Trait ResultAdapter

pub trait ResultAdapter {
    type Ok;
    type Err;
}

Helper trait for macro-generated code.

This trait allows macros to refer to the Ok and Err types of a Result that is behind a type alias. This is currently required because the macros for properties expect a Result return value, but the macro-generated receive_ functions need to refer to the actual type without the associated error.

Required Associated Types

type Ok

type Err