Module zbus::blocking

The blocking API.

This module hosts all our blocking API. All the types under this module are thin wrappers around the corresponding asynchronous types. Most of the method calls are simply calling their asynchronous counterparts on the underlying types and use async_io::block_on to turn them into blocking calls.

Caveats

Since methods provided by these types run their own little runtime (block_on), you must not use them in async contexts because of the infamous async sandwich footgun. This is an especially important fact to keep in mind for crate::dbus_interface. While dbus_interface allows non-async methods for convenience, these methods are called from an async context. The blocking crate provides an easy way around this problem though.

Modules

fdo

D-Bus standard interfaces.

Structs

Connection

A blocking wrapper of zbus::Connection.

ConnectionBuilder

A builder for zbus::blocking::Connection.

InterfaceRef

Wrapper over an interface, along with its corresponding SignalContext instance. A reference to the underlying interface may be obtained via InterfaceRef::get and InterfaceRef::get_mut.

MessageIterator

A blocking wrapper of crate::MessageStream.

ObjectServer

A blocking wrapper of crate::ObjectServer.

OwnerChangedIterator

An std::iter::Iterator implementation that yields owner change notifications.

PropertyChanged

A property changed event.

PropertyIterator

An std::iter::Iterator implementation that yields property change notifications.

Proxy

A blocking wrapper of crate::Proxy.

ProxyBuilder

Builder for proxies.

SignalIterator

An std::iter::Iterator implementation that yields signal messages.

Traits

ProxyDefault

Trait for the default associated values of a proxy.