Module zbus::fdo

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

ConnectionCredentials

Credentials of a process connected to a bus server.

DBusProxy

Proxy for the org.freedesktop.DBus interface.

InterfacesAdded

A InterfacesAdded signal.

InterfacesAddedArgs

InterfacesAdded signal arguments.

InterfacesAddedStream

A Stream implementation that yields InterfacesAdded signals.

InterfacesRemoved

A InterfacesRemoved signal.

InterfacesRemovedArgs

InterfacesRemoved signal arguments.

InterfacesRemovedStream

A Stream implementation that yields InterfacesRemoved signals.

IntrospectableProxy

Proxy for the org.freedesktop.DBus.Introspectable interface.

MonitoringProxy

Proxy for the org.freedesktop.DBus.Monitoring interface.

NameAcquired

A NameAcquired signal.

NameAcquiredArgs

NameAcquired signal arguments.

NameAcquiredStream

A Stream implementation that yields NameAcquired signals.

NameLost

A NameLost signal.

NameLostArgs

NameLost signal arguments.

NameLostStream

A Stream implementation that yields NameLost signals.

NameOwnerChanged

A NameOwnerChanged signal.

NameOwnerChangedArgs

NameOwnerChanged signal arguments.

NameOwnerChangedStream

A Stream implementation that yields NameOwnerChanged signals.

ObjectManager

Service-side Object Manager interface implementation.

ObjectManagerProxy

Proxy for the org.freedesktop.DBus.ObjectManager interface.

PeerProxy

Proxy for the org.freedesktop.DBus.Peer interface.

Properties

Server-side implementation for the org.freedesktop.DBus.Properties interface. This interface is implemented automatically for any object registered to the ObjectServer.

PropertiesChanged

A PropertiesChanged signal.

PropertiesChangedArgs

PropertiesChanged signal arguments.

PropertiesChangedStream

A Stream implementation that yields PropertiesChanged signals.

PropertiesProxy

Proxy for the org.freedesktop.DBus.Properties interface.

StatsProxy

Proxy for the org.freedesktop.DBus.Stats interface.

Enums

Error

Errors from https://gitlab.freedesktop.org/dbus/dbus/-/blob/master/dbus/dbus-protocol.h

ReleaseNameReply

The return code of the release_name method.

RequestNameFlags

The flags used by the bus request_name method.

RequestNameReply

The return code of the request_name method.

Type Definitions

ManagedObjects

The type returned by the ObjectManagerProxy::get_managed_objects method.

Result

Alias for a Result with the error type zbus::fdo::Error.