Module zbus::fdo

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

DBusProxy
Proxy for the org.freedesktop.DBus interface.

IntrospectableProxy
Proxy for the org.freedesktop.DBus.Introspectable interface.

MonitoringProxy
Proxy for the org.freedesktop.DBus.Monitoring interface.

ObjectManagerProxy
Proxy for the org.freedesktop.DBus.ObjectManager interface.

PeerProxy
Proxy for the org.freedesktop.DBus.Peer interface.

PropertiesProxy
Proxy for the org.freedesktop.DBus.Properties interface.

StatsProxy
Proxy for the org.freedesktop.DBus.Stats interface.

Enums

Error
Errors from https://gitlab.freedesktop.org/dbus/dbus/-/blob/master/dbus/dbus-protocol.h

RequestNameFlags
The flags used by the bus request_name method.

RequestNameReply
The return code of the request_name method.

Type Definitions

Result
Alias for a Result with the error type zbus::fdo::Error.