Enum zbus::fdo::Error

pub enum Error {

    ZBus(Error),
    Failed(String),
    NoMemory(String),
    ServiceUnknown(String),
    NameHasNoOwner(String),
    NoReply(String),
    IOError(String),
    BadAddress(String),
    NotSupported(String),
    LimitsExceeded(String),
    AccessDenied(String),
    AuthFailed(String),
    NoServer(String),
    Timeout(String),
    NoNetwork(String),
    AddressInUse(String),
    Disconnected(String),
    InvalidArgs(String),
    FileNotFound(String),
    FileExists(String),
    UnknownMethod(String),
    UnknownObject(String),
    UnknownInterface(String),
    UnknownProperty(String),
    PropertyReadOnly(String),
    TimedOut(String),
    MatchRuleNotFound(String),
    MatchRuleInvalid(String),
    SpawnExecFailed(String),
    SpawnForkFailed(String),
    SpawnChildExited(String),
    SpawnChildSignaled(String),
    SpawnFailed(String),
    SpawnFailedToSetup(String),
    SpawnConfigInvalid(String),
    SpawnServiceNotValid(String),
    SpawnServiceNotFound(String),
    SpawnPermissionsInvalid(String),
    SpawnFileInvalid(String),
    SpawnNoMemory(String),
    UnixProcessIdUnknown(String),
    InvalidSignature(String),
    InvalidFileContent(String),
    SELinuxSecurityContextUnknown(String),
    AdtAuditDataUnknown(String),
    ObjectPathInUse(String),
    InconsistentMessage(String),
    InteractiveAuthorizationRequired(String),
    NotContainer(String),

}

Errors from https://gitlab.freedesktop.org/dbus/dbus/-/blob/master/dbus/dbus-protocol.h

Variants

ZBus(Error)

Unknown or fall-through ZBus error.

Failed(String)

A generic error; “something went wrong” - see the error message for more.

NoMemory(String)

There was not enough memory to complete an operation.

ServiceUnknown(String)

The bus doesn’t know how to launch a service to supply the bus name you wanted.

NameHasNoOwner(String)

The bus name you referenced doesn’t exist (i.e. no application owns it).

NoReply(String)

No reply to a message expecting one, usually means a timeout occurred.

IOError(String)

Something went wrong reading or writing to a socket, for example.

BadAddress(String)

A D-Bus bus address was malformed.

NotSupported(String)

Requested operation isn’t supported (like ENOSYS on UNIX).

LimitsExceeded(String)

Some limited resource is exhausted.

AccessDenied(String)

Security restrictions don’t allow doing what you’re trying to do.

AuthFailed(String)

Authentication didn’t work.

NoServer(String)

Unable to connect to server (probably caused by ECONNREFUSED on a socket).

Timeout(String)

Certain timeout errors, possibly ETIMEDOUT on a socket. Note that TimedOut is used for message reply timeouts.

NoNetwork(String)

No network access (probably ENETUNREACH on a socket).

AddressInUse(String)

Can’t bind a socket since its address is in use (i.e. EADDRINUSE).

Disconnected(String)

The connection is disconnected and you’re trying to use it.

InvalidArgs(String)

Invalid arguments passed to a method call.

FileNotFound(String)

Missing file.

FileExists(String)

Existing file and the operation you’re using does not silently overwrite.

UnknownMethod(String)

Method name you invoked isn’t known by the object you invoked it on.

UnknownObject(String)

Object you invoked a method on isn’t known.

UnknownInterface(String)

Interface you invoked a method on isn’t known by the object.

UnknownProperty(String)

Property you tried to access isn’t known by the object.

PropertyReadOnly(String)

Property you tried to set is read-only.

TimedOut(String)

Certain timeout errors, e.g. while starting a service.

MatchRuleNotFound(String)

Tried to remove or modify a match rule that didn’t exist.

MatchRuleInvalid(String)

The match rule isn’t syntactically valid.

SpawnExecFailed(String)

While starting a new process, the exec() call failed.

SpawnForkFailed(String)

While starting a new process, the fork() call failed.

SpawnChildExited(String)

While starting a new process, the child exited with a status code.

SpawnChildSignaled(String)

While starting a new process, the child exited on a signal.

SpawnFailed(String)

While starting a new process, something went wrong.

SpawnFailedToSetup(String)

We failed to setup the environment correctly.

SpawnConfigInvalid(String)

We failed to setup the config parser correctly.

SpawnServiceNotValid(String)

Bus name was not valid.

SpawnServiceNotFound(String)

Service file not found in system-services directory.

SpawnPermissionsInvalid(String)

Permissions are incorrect on the setuid helper.

SpawnFileInvalid(String)

Service file invalid (Name, User or Exec missing).

SpawnNoMemory(String)

There was not enough memory to complete the operation.

UnixProcessIdUnknown(String)

Tried to get a UNIX process ID and it wasn’t available.

InvalidSignature(String)

A type signature is not valid.

InvalidFileContent(String)

A file contains invalid syntax or is otherwise broken.

SELinuxSecurityContextUnknown(String)

Asked for SELinux security context and it wasn’t available.

AdtAuditDataUnknown(String)

Asked for ADT audit data and it wasn’t available.

ObjectPathInUse(String)

There’s already an object with the requested object path.

InconsistentMessage(String)

The message meta data does not match the payload. e.g. expected number of file descriptors were not sent over the socket this message was received on.

InteractiveAuthorizationRequired(String)

The message is not allowed without performing interactive authorization, but could have succeeded if an interactive authorization step was allowed.

NotContainer(String)

The connection is not from a container, or the specified container instance does not exist.