zbus::fdo

Module object_manager

D-Bus standard interfaces.

The D-Bus specification defines the message bus messages and some standard interfaces that may be useful across various D-Bus applications. This module provides their proxy.

Structs

InterfacesAdded

A InterfacesAdded signal.

InterfacesAddedArgs

InterfacesAdded signal arguments.

InterfacesAddedStream

A Stream implementation that yields InterfacesAdded signals.

InterfacesRemoved

A InterfacesRemoved signal.

InterfacesRemovedArgs

InterfacesRemoved signal arguments.

InterfacesRemovedStream

A Stream implementation that yields InterfacesRemoved signals.

ObjectManager

Service-side Object Manager interface implementation.

ObjectManagerProxy

Proxy for the org.freedesktop.DBus.ObjectManager interface.

Traits

ObjectManagerSignals

Trait providing all signal emission methods for ObjectManager.

Type Aliases

ManagedObjects

The type returned by the ObjectManagerProxy::get_managed_objects method.