zbus::fdo::stats

Struct ConnectionStats

pub struct ConnectionStats {

    pub(crate) serial: Option<u32>,
    pub(crate) unique_name: Option<OwnedUniqueName>,
    pub(crate) match_rules: Option<u32>,
    pub(crate) peak_match_rules: Option<u32>,
    pub(crate) bus_names: Option<u32>,
    pub(crate) peak_bus_names: Option<u32>,
    pub(crate) incoming_messages: Option<u32>,
    pub(crate) outgoing_messages: Option<u32>,
    pub(crate) incoming_bytes: Option<u32>,
    pub(crate) outgoing_bytes: Option<u32>,
    pub(crate) incoming_fds: Option<u32>,
    pub(crate) outgoing_fds: Option<u32>,
    pub(crate) peak_incoming_messages: Option<u32>,
    pub(crate) peak_outgoing_messages: Option<u32>,
    pub(crate) peak_incoming_bytes: Option<u32>,
    pub(crate) peak_outgoing_bytes: Option<u32>,
    pub(crate) peak_incoming_fds: Option<u32>,
    pub(crate) peak_outgoing_fds: Option<u32>,
    pub(crate) rest: HashMap<String, OwnedValue>,

}

The stats returned by the StatsProxy::get_connection_stats method.

Fields
serial: Option<u32>unique_name: Option<OwnedUniqueName>match_rules: Option<u32>peak_match_rules: Option<u32>bus_names: Option<u32>peak_bus_names: Option<u32>incoming_messages: Option<u32>outgoing_messages: Option<u32>incoming_bytes: Option<u32>outgoing_bytes: Option<u32>incoming_fds: Option<u32>outgoing_fds: Option<u32>peak_incoming_messages: Option<u32>peak_outgoing_messages: Option<u32>peak_incoming_bytes: Option<u32>peak_outgoing_bytes: Option<u32>peak_incoming_fds: Option<u32>peak_outgoing_fds: Option<u32>rest: HashMap<String, OwnedValue>
Implementations

impl ConnectionStats

pub fn serial(&self) -> Option<u32>

A serial number which is incremented with each call to the GetConnectionStats method.

pub fn unique_name(&self) -> Option<&OwnedUniqueName>

The unique name of the connection.

pub fn match_rules(&self) -> Option<u32>

The total number of match rules that are currently in use by this connection.

pub fn peak_match_rules(&self) -> Option<u32>

The largest total number of match rules that have been in use by this connection at any one time.

pub fn bus_names(&self) -> Option<u32>

The total number of bus names that are currently in use by this connection.

pub fn peak_bus_names(&self) -> Option<u32>

The largest total number of bus names that have been in use by this connection at any one time.

pub fn incoming_messages(&self) -> Option<u32>

The total number of messages received by this connection.

pub fn outgoing_messages(&self) -> Option<u32>

The total number of messages sent by this connection.

pub fn incoming_bytes(&self) -> Option<u32>

The total number of bytes received by this connection.

pub fn outgoing_bytes(&self) -> Option<u32>

The total number of bytes sent by this connection.

pub fn incoming_fds(&self) -> Option<u32>

The total number of file descriptors received by this connection.

pub fn outgoing_fds(&self) -> Option<u32>

The total number of file descriptors sent by this connection.

pub fn peak_incoming_messages(&self) -> Option<u32>

The largest total number of messages that have been in use by this connection at any one time.

pub fn peak_outgoing_messages(&self) -> Option<u32>

The largest total number of messages that have been in use by this connection at any one time.

pub fn peak_incoming_bytes(&self) -> Option<u32>

The largest total number of bytes that have been in use by this connection at any one time.

pub fn peak_outgoing_bytes(&self) -> Option<u32>

The largest total number of bytes that have been in use by this connection at any one time.

pub fn peak_incoming_fds(&self) -> Option<u32>

The largest total number of file descriptors that have been in use by this connection at any one time.

pub fn peak_outgoing_fds(&self) -> Option<u32>

The largest total number of file descriptors that have been in use by this connection at any one time.

pub fn rest(&self) -> &HashMap<String, OwnedValue>

The rest of the statistics, that are not defined by the D-Bus specificiation and hence specific to individual D-Bus broker implementations.