Struct zbus::fdo::DBusProxy

pub struct DBusProxy<'c>(_);

Proxy for the org.freedesktop.DBus interface.

Implementations

impl<'c> DBusProxy<'c>

pub fn new(conn: &Connection) -> Result<Self>

Creates a new proxy with the default service & path.

pub fn new_for(
conn: &Connection,
destination: &'c str,
path: &'c str
) -> Result<Self>

Creates a new proxy for the given destination and path.

pub fn new_for_owned(
conn: Connection,
destination: String,
path: String
) -> Result<Self>

Same as new_for but takes ownership of the passed arguments.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy

The reference to the underlying zbus::Proxy.

pub fn add_match(&self, rule: &str) -> Result<()>

Adds a match rule to match messages going through the message bus

pub fn get_adt_audit_session_data(&self, bus_name: &str) -> Result<Vec<u8>>

Returns auditing data used by Solaris ADT, in an unspecified binary format.

pub fn get_connection_credentials(
&self,
bus_name: &str
) -> Result<HashMap<String, OwnedValue>>

Returns as many credentials as possible for the process connected to the server.

pub fn get_connection_selinux_security_context(
&self,
bus_name: &str
) -> Result<Vec<u8>>

Returns the security context used by SELinux, in an unspecified format.

pub fn get_connection_unix_process_id(&self, bus_name: &str) -> Result<u32>

Returns the Unix process ID of the process connected to the server.

pub fn get_connection_unix_user(&self, bus_name: &str) -> Result<u32>

Returns the Unix user ID of the process connected to the server.

pub fn get_id(&self) -> Result<String>

Gets the unique ID of the bus.

pub fn get_name_owner(&self, name: &str) -> Result<String>

Returns the unique connection name of the primary owner of the name given.

pub fn hello(&self) -> Result<String>

Returns the unique name assigned to the connection.

pub fn list_activatable_names(&self) -> Result<Vec<String>>

Returns a list of all names that can be activated on the bus.

pub fn list_names(&self) -> Result<Vec<String>>

Returns a list of all currently-owned names on the bus.

pub fn list_queued_owners(&self, name: &str) -> Result<Vec<String>>

List the connections currently queued for a bus name.

pub fn name_has_owner(&self, name: &str) -> Result<bool>

Checks if the specified name exists (currently has an owner).

pub fn release_name(&self, name: &str) -> Result<()>

Ask the message bus to release the method caller's claim to the given name.

pub fn reload_config(&self) -> Result<()>

Reload server configuration.

pub fn remove_match(&self, rule: &str) -> Result<()>

Removes the first rule that matches.

pub fn request_name(
&self,
name: &str,
flags: BitFlags<RequestNameFlags>
) -> Result<RequestNameReply>

Ask the message bus to assign the given name to the method caller.

pub fn start_service_by_name(&self, name: &str, flags: u32) -> Result<u32>

Tries to launch the executable associated with a name (service activation), as an explicit request.

pub fn update_activation_environment(
&self,
environment: HashMap<&str, &str>
) -> Result<()>

This method adds to or modifies that environment when activating services.

pub fn features(&self) -> Result<Vec<String>>

This signal indicates that the owner of a name has changed. It's also the signal to use to detect the appearance of new names on the bus. This signal is sent to a specific application when it loses ownership of a name. This signal is sent to a specific application when it gains ownership of a name. This property lists abstract “features” provided by the message bus, and can be used by clients to detect the capabilities of the message bus with which they are communicating.

pub fn interfaces(&self) -> Result<Vec<String>>

This property lists interfaces provided by the /org/freedesktop/DBus object, and can be used by clients to detect the capabilities of the message bus with which they are communicating. Unlike the standard Introspectable interface, querying this property does ot require parsing XML. This property was added in version 1.11.x of the reference implementation of the message bus.

The standard org.freedesktop.DBus and org.freedesktop.DBus.Properties interfaces are not included in the value of this property, because their presence can be inferred from the fact that a method call on org.freedesktop.DBus.Properties asking for properties of org.freedesktop.DBus was successful. The standard org.freedesktop.DBus.Peer and org.freedesktop.DBus.Introspectable interfaces are not included in the value of this property either, because they do not indicate features of the message bus implementation.

pub fn introspect(&self) -> Result<String>