Module zbus::object_server

The object server API.

Structs

InterfaceDeref

Opaque structure that derefs to an Interface type.

InterfaceDerefMut

Opaque structure that mutably derefs to an Interface type.

InterfaceRef

Wrapper over an interface, along with its corresponding SignalContext instance. A reference to the underlying interface may be obtained via InterfaceRef::get and InterfaceRef::get_mut.

ObjectServer

An object server, holding server-side D-Bus objects & interfaces.

ResponseDispatchNotifier

A response wrapper that notifies after response has been sent.

SignalContext

A signal emission context.

Enums

DispatchResult

A helper type returned by Interface callbacks.

Traits

Interface

The trait is used to dispatch messages to an interface instance.