zbus
Module proxy

The client-side proxy API.

Structs

Builder

Builder for proxies.

OwnerChangedStream

A stream::Stream implementation that yields UniqueName when the bus owner changes.

PropertyChanged

A property changed event.

PropertyStream

A stream::Stream implementation that yields property change notifications.

Proxy

A client-side interface proxy.

SignalStream

A stream::Stream implementation that yields signal messages.

Enums

CacheProperties

The properties caching mode.

MethodFlags

Flags to use with Proxy::call_with_flags.

Traits

ProxyDefault

Trait for the default associated values of a proxy.

ProxyImpl

This trait is implemented by all async proxies, which are generated with the dbus_proxy macro.